requirements.txt
aiohappyeyeballs==2.4.8
aiohttp==3.11.13
aiosignal==1.3.2
archiveis==0.0.9
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
discord.py==2.5.1
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.3.0
requests==2.32.3
urllib3==2.3.0
yarl==1.18.3
